socket-server/requirements.txt
aio-pika==9.5.5
aioredis==2.0.1
aiormq==6.8.1
annotated-types==0.7.0
async-timeout==5.0.1
asyncpg==0.30.0
bidict==0.23.1
click==8.1.8
colorama==0.4.6
dotenv==0.9.9
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
multidict==6.4.3
pamqp==3.3.0
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
python-engineio==4.12.0
python-socketio==5.13.0
simple-websocket==1.1.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
wsproto==1.2.0
yarl==1.20.0
